+++ .eslintrc.json
@@ -6,7 +6,6 @@
   },
   "rules": {
     "no-plusplus": "off",
-    "indent": ["error", 4],
     "react/jsx-filename-extension": [
       1,
       {
